split skillsets client
